# `Node`: Validate nested attribute and extra keys