@mbercx 2023-04-15
I've done some benchmarks to get an idea of the impact on performance. For an independent recursive function:

```python
def validate_dict(dictionary):
    for key, value in dictionary.items():
        if not isinstance(key, str):
            raise TypeError(f'Found non-string key `{key}` in dictionary with type `{type(key)}`.')
        if isinstance(value, dict):
            validate_dict(value)
```

and the following dictionaries:

```python
d1 = {str(top): value for top, value in zip(range(100), range(100))}
d2 = {
    str(top): {str(mid): value for mid, value in zip(range(100), range(100))}
    for top in range(100)
}
d3 = {
    str(top): {
        str(mid): {str(deep): value for deep, value in zip(range(100), range(100))}
        for mid in range(100)
    }
    for top in range(100)
}
```

I get the following timings for the `validate_dict` function:

```
d1: 6.45 µs ± 99.8 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
d2: 665 µs ± 9.86 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
d3: 67.6 ms ± 647 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

So, pretty linear scaling vs the number of keys, which is sort of expected.

When I run

```python
%%timeit -n 10

orm.Dict(<dictionary>).store()
```

for the three dictionaries, I get the following comparison of performance before and after adding the validation:

```
d1:
- before: 14.9 ms ± 2.74 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
- after: 12.5 ms ± 2.39 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

```
d2:
- before: 106 ms ± 2.64 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
- after: 107 ms ± 1.89 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

```
d3:
- before: 7.79 s ± 42 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
- after: 8.05 s ± 22.2 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

So there might be some _small_ influence on the performance supermassive dictionary, but nothing critical.