@mbercx 2023-05-29
Just a thought: this change is actually backwards-incompatible. It's possible some plugins rely on automatic casting of the nested `Dict` keys as strings, and their code might break.