@sphuber 2023-07-11
> Just a thought: this change is actually backwards-incompatible. It's possible some plugins rely on automatic casting of the nested `Dict` keys as strings, and their code might break.

Coming back to this: this is the reason I didn't include this in v2.4. I think we can simply try/catch the added validation and for now just add a `warn.warn_deprecation` message saying that this behavior will start to except starting from AiiDA v3.0  Then we can merge this PR. Do you see a problem with that approach @mbercx ?