@sphuber 2023-09-05
Note that I came across an example that is currently problematic but would actually be fixed by this PR. If you construct a `Dict` with a tuple as a _nested_ key, you can create the node just fine, but when storing, the code will except on the database insertion level. This can leave the Sqlalchemy session in a broken state. Adding this validation before storing would solve this.